Fixed a potential missed bus issue
If the passenger was sold a ticket, but then the bus departed there was the potential for a customer
to be stuck waiting for the boarding condition forever. Attempting to fix this by halting ticket sales when the bus is readying for departure and then making the bus wait until all passengers have boarded.