Removed incorrect sentence
Said that the signalled calls form a stack that is pushed to and popped from as processes signal and wait/end. However, in reality this is not the case because the next semaphore is serviced in a FIFO order. So if P1 signals P2 and P2 signals P3 when P3 ends P1 will return and then after P1 P2 will return.